Connected to scope, created new main to interface w scope, added visa files and altered makefile to include visa libraries

+++ includes/dataprocessing.h
@@ -1,10 +1,10 @@
 #ifndef __DATAPROCESSING_H
 #define __DATAPROCESSING_H
 
-float mean(float num[], int n);
-float stdev(float num[], float mean, int n);
-float rms(float num[], int n);
-float Amp(float num[], int n);
-float smooth(double *d, int n, int w, double* end);
+double mean(double num[], int n);
+double stdev(double num[], double mean, int n);
+double rms(double num[], int n);
+double Amp(double num[], int n);
+double smooth(double *d, int n, int w, double* end);
 
 #endif
